Rename ssf metric extraction sink package to ssfmetrics
